+++ hyper_chat/api/apis/api_chatroom.py
@@ -17,13 +17,19 @@ class ChatRoomList(APIView):
         return Response(serializers.data, status=status.HTTP_200_OK)
 
     def post(self, request):
-        chatroom_filter = ChatRoom.objects.filter(owner_id=request.data.get('owner_id'),
+        chatroom_filter = ChatRoom.objects.filter(owner_id=request.user.id,
                                                   opponent_id=request.data.get('opponent_id'))
-
+        if chatroom_filter.count() > 0:
+            return Response({'Error': 'Already Exist ChatRoom'}, status=status.HTTP_400_BAD_REQUEST)
+        
+        chatroom_filter = ChatRoom.objects.filter(owner_id=request.data.get('opponent_id'),
+                                                  opponent_id=request.user.id)
         if chatroom_filter.count() > 0:
             return Response({'Error': 'Already Exist ChatRoom'}, status=status.HTTP_400_BAD_REQUEST)
 
-        serializer = ChatRoomSerializer(data=request.data)
+        data = request.data
+        data['owner_id'] = request.user.id
+        serializer = ChatRoomSerializer(data=data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
@@ -36,10 +42,16 @@ class ChatRoomDetail(APIView):
 
     def get(self, request, chatroom_id):
         chatroom = ChatRoom.objects.get(id=chatroom_id)
+        if not (chatroom.owner.id == request.user.id or chatroom.opponent.id == request.user.id):
+            return Response({'Error': 'Unauthorized Error'}, status=status.HTTP_401_UNAUTHORIZED)
+
         serializer = ChatRoomSerializer(chatroom)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
     def delete(self, request, chatroom_id):
         chatroom = ChatRoom.objects.get(id=chatroom_id)
+        if not (chatroom.owner.id == request.user.id or chatroom.opponent.id == request.user.id):
+            return Response({'Error': 'Unauthorized Error'}, status=status.HTTP_401_UNAUTHORIZED)
+
         chatroom.delete()
         return Response({'message': 'Success Delete Chatroom'}, status=status.HTTP_200_OK)
